+++ create_dataframe.py
@@ -12,7 +12,7 @@ with open(sys.argv[1]) as f_:
     config = yaml.load(f_, Loader=yaml.FullLoader)
 
 def create_dataframe(fl, fpl, config, slice):
-    if config['merge']['paired_End'] and not config['general']['already_assembled']:
+    if config['general']['paired_End'] and not config['general']['already_assembled']:
         df = pd.DataFrame(columns=['sample', 'unit', 'fq1', 'fq2'],
             index =range(int(len(fl)/2)), dtype=str)
         i, j = (0, 0)
@@ -37,7 +37,8 @@ def create_dataframe(fl, fpl, config, slice):
 
 if __name__ == '__main__':
     if not config['general']['already_assembled']:
-        file_path_list = sorted(glob('demultiplexed/*.gz'))
+        file_path_list = ['demultiplexed/' + name.split('/')[-1] for name in
+                          sorted(glob(config['general']['filename'] + '/*.gz'))]
         file_list = sorted([file_.split('/')[-1] for file_ 
                     in file_path_list])
         slice = -3 # Remove the .gz extension from the file paths.