+++ pipeline.sh
@@ -6,11 +6,10 @@ read varname
 env_loc=$(conda info --base)/etc/profile.d/conda.sh
 
 source $env_loc
-conda activate snakemake
-python demultiplexing.py "$varname"
+conda activate natrix_repair
 
 varname+=".yaml"
 cores=$(grep "cores : " $varname | awk '{print $3}')
 python create_dataframe.py "$varname"
 
-screen -S $varname bash -c "source $env_loc;conda activate snakemake;snakemake --use-conda --cores $cores --configfile $varname; exec sh"
+screen -S $varname bash -c "source $env_loc;conda activate natrix_repair;snakemake --use-conda --cores $cores --configfile $varname; exec sh"